Re-organize under src directory.
