+++ src/tests.py
@@ -12,7 +12,7 @@ in the folder where this file exists:
 """
 import unittest
 
-from server_logic import avoid_my_neck
+import logic
 
 
 class AvoidNeckTest(unittest.TestCase):
@@ -24,12 +24,11 @@ class AvoidNeckTest(unittest.TestCase):
         single place, and thus all directions are valid.
         """
         # Arrange
-        test_head = {"x": 5, "y": 5}
         test_body = [{"x": 5, "y": 5}, {"x": 5, "y": 5}, {"x": 5, "y": 5}]
         possible_moves = ["up", "down", "left", "right"]
 
         # Act
-        result_moves = avoid_my_neck(test_head, test_body, possible_moves)
+        result_moves = logic._avoid_my_neck(test_body, possible_moves)
 
         # Assert
         self.assertEqual(len(result_moves), 4)
@@ -37,13 +36,12 @@ class AvoidNeckTest(unittest.TestCase):
 
     def test_avoid_neck_left(self):
         # Arrange
-        test_head = {"x": 5, "y": 5}
         test_body = [{"x": 5, "y": 5}, {"x": 4, "y": 5}, {"x": 3, "y": 5}]
         possible_moves = ["up", "down", "left", "right"]
         expected = ["up", "down", "right"]
 
         # Act
-        result_moves = avoid_my_neck(test_head, test_body, possible_moves)
+        result_moves = logic._avoid_my_neck(test_body, possible_moves)
 
         # Assert
         self.assertEqual(len(result_moves), 3)
@@ -51,13 +49,12 @@ class AvoidNeckTest(unittest.TestCase):
 
     def test_avoid_neck_right(self):
         # Arrange
-        test_head = {"x": 5, "y": 5}
         test_body = [{"x": 5, "y": 5}, {"x": 6, "y": 5}, {"x": 7, "y": 5}]
         possible_moves = ["up", "down", "left", "right"]
         expected = ["up", "down", "left"]
 
         # Act
-        result_moves = avoid_my_neck(test_head, test_body, possible_moves)
+        result_moves = logic._avoid_my_neck(test_body, possible_moves)
 
         # Assert
         self.assertEqual(len(result_moves), 3)
@@ -65,13 +62,12 @@ class AvoidNeckTest(unittest.TestCase):
 
     def test_avoid_neck_up(self):
         # Arrange
-        test_head = {"x": 5, "y": 5}
         test_body = [{"x": 5, "y": 5}, {"x": 5, "y": 6}, {"x": 5, "y": 7}]
         possible_moves = ["up", "down", "left", "right"]
         expected = ["down", "left", "right"]
 
         # Act
-        result_moves = avoid_my_neck(test_head, test_body, possible_moves)
+        result_moves = logic._avoid_my_neck(test_body, possible_moves)
 
         # Assert
         self.assertEqual(len(result_moves), 3)
@@ -79,13 +75,12 @@ class AvoidNeckTest(unittest.TestCase):
 
     def test_avoid_neck_down(self):
         # Arrange
-        test_head = {"x": 5, "y": 5}
         test_body = [{"x": 5, "y": 5}, {"x": 5, "y": 4}, {"x": 5, "y": 3}]
         possible_moves = ["up", "down", "left", "right"]
         expected = ["up", "left", "right"]
 
         # Act
-        result_moves = avoid_my_neck(test_head, test_body, possible_moves)
+        result_moves = logic._avoid_my_neck(test_body, possible_moves)
 
         # Assert
         self.assertEqual(len(result_moves), 3)
